+++ src/test/java/com/braintreegateway/integrationtest/TransactionIT.java
@@ -1104,7 +1104,37 @@ public class TransactionIT extends IntegrationTest implements MerchantAccountTes
                 eciFlag("02").
                 cavv("some_cavv").
                 xid("some_xid").
+                authenticationResponse("Y").
+                directoryResponse("Y").
+                cavvAlgorithm("2").
+                threeDSecureVersion("1.0.2").
                 done();
+
+        Result<Transaction> result = gateway.transaction().sale(request);
+        assertTrue(result.isSuccess());
+
+        Transaction transaction = result.getTarget();
+        assertEquals(Transaction.Status.AUTHORIZED, transaction.getStatus());
+    }
+
+    @Test
+    public void saleWithThreeDSecurePassThruVersion2() {
+        TransactionRequest request = new TransactionRequest().
+            merchantAccountId(THREE_D_SECURE_MERCHANT_ACCOUNT_ID).
+            amount(TransactionAmount.AUTHORIZE.amount).
+            creditCard().
+            number(CreditCardNumber.VISA.number).
+            expirationDate("05/2009").
+            done().
+            threeDSecurePassThru().
+            eciFlag("02").
+            cavv("some_cavv").
+            authenticationResponse("Y").
+            directoryResponse("Y").
+            cavvAlgorithm("2").
+            threeDSecureVersion("2.0.1").
+            dsTransactionId("some_ds_transaction_id").
+            done();
 
         Result<Transaction> result = gateway.transaction().sale(request);
         assertTrue(result.isSuccess());
@@ -1131,6 +1161,7 @@ public class TransactionIT extends IntegrationTest implements MerchantAccountTes
                 authenticationResponse("Y").
                 directoryResponse("Y").
                 cavvAlgorithm("2").
+                threeDSecureVersion("1.0.2").
                 done();
 
         Result<Transaction> result = gateway.transaction().sale(request);